+++ song-service/src/main/java/io/github/bohdanrakov/dtos/MP3MetadataResponse.java
@@ -1,5 +1,5 @@
 package io.github.bohdanrakov.dtos;
 
-public record MP3MetadataResponse(Integer id, String name, String artist, String album, String duration, String year) {
+public record MP3MetadataResponse(Long id, String name, String artist, String album, String duration, String year) {
 
 }
